Refactor JdbcService.
