+++ src/main/scala/mirroring/handlers/ChangeTrackingHandler.scala
@@ -21,14 +21,14 @@ import mirroring.Config
 import java.sql.DriverManager
 import org.apache.spark.sql.DataFrame
 import mirroring.services.SparkService.spark
-import mirroring.services.databases.{DbService, JdbcService}
+import mirroring.services.databases.JdbcService
 import wvlet.log.LogSupport
 import mirroring.builders._
 
 class ChangeTrackingHandler(config: Config) extends LogSupport {
 
   private val jdbcContext                  = config.getJdbcContext
-  private val jdbcService: DbService       = new JdbcService(jdbcContext)
+  private val jdbcService: JdbcService     = new JdbcService(jdbcContext)
 
   private def getChangeTrackingVersionCustom(jdbcUrl: String,
                                              query: String,
@@ -127,9 +127,9 @@ class ChangeTrackingHandler(config: Config) extends LogSupport {
 
   def getChangeTrackingVersionDefault(query: String, config: Config): BigInt = {
 
-    val jdbcContext            = config.getJdbcContext
-    val jdbcService: DbService = new JdbcService(jdbcContext)
-    val jdbcDF: DataFrame      = jdbcService.loadData(query).cache()
+    val jdbcContext              = config.getJdbcContext
+    val jdbcService: JdbcService = new JdbcService(jdbcContext)
+    val jdbcDF: DataFrame        = jdbcService.loadData(query).cache()
 
     var version: BigInt = BigInt(0)
